requirements.txt
streamlit
yfinance
matplotlib
textblob
python-dateutil
numpy_financial